Refactor process.go for platform specific
